+++ src/Traits/DatabaseTrait.php
@@ -2,8 +2,6 @@
 
 namespace Rougin\Wildfire\Traits;
 
-use Rougin\SparkPlug\Instance;
-
 use Rougin\Wildfire\Helpers\DescribeHelper;
 
 /**
@@ -31,26 +29,50 @@ trait DatabaseTrait
 
     /**
      * Sets the database class.
+     * NOTE: To be removed in v1.0.0. Use $this->database instead.
      *
      * @param  \CI_DB_query_builder $database
      * @return self
      */
     public function setDatabase($database)
     {
+        return $this->database($database);
+    }
+
+    /**
+     * Sets the database instance.
+     *
+     * @param  \CI_DB_query_builder $database
+     * @return self
+     */
+    public function database($database)
+    {
         $this->db = $database;
 
-        $this->describe = DescribeHelper::createInstance($database);
+        $this->describe = DescribeHelper::make($database);
 
         return $this;
     }
 
     /**
      * Sets the query result.
+     * NOTE: To be removed in v1.0.0. Use $this->query instead.
      *
      * @param  \CI_DB_result $query
      * @return self
      */
     public function setQuery($query)
+    {
+        return $this->query($query);
+    }
+
+    /**
+     * Sets the query result instance.
+     *
+     * @param  \CI_DB_result $query
+     * @return self
+     */
+    public function query($query)
     {
         $this->query = $query;